Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,8 +3,8 @@
 
 
 int main() {
-    int n,m,s;
-    scanf("%d %d %d",&n,&m,&s);
+    float n,m,s;
+    scanf("%f %f %f",&n,&m,&s);
     float avg=(n+m+s)/3;
     printf("Average: %.2f\n",avg);
     return 0;
